implemeted ai service to the backend and fixed some report issues

+++ BinMaps.Shared/DTOs/AIResultDTO.cs
@@ -6,7 +6,7 @@ namespace BinMaps.Shared.DTOs
 {
     public class AIResultDto
     {
-        public int FillPercentage { get; set; }
+        public double FillPercentage { get; set; }
         public bool FireDetected { get; set; }
         public int Confidence { get; set; }
     }
